listen to signals for debugging and profiling
Some interesting findings when running profiles:

There's a good performance improvement by compiling with Go tip due to
golang/go@6fd82d83d697edebbe826a028e752c96f195111f commit. The total GC
pause time in a 60s run dropped from ~4s to ~50ms.

Setting `MaxIdleConnsPerHost` in the transport has a *huge* performance
impact. If this number is big enough to completely prevent the server
from closing and opening connections on each request the throughput is
awesome.

The downside is that this connections will remain open for each host
indeterminately and if we have a big number of hosts we're sending
requests to it will be easy to exhaust all available ports with idle
connections.